make log files have unique file names
This seemed to be useful to do when re-rerunning the gating, otherwise the original log file blocked a new one being created. I think, however, that the new log files will not be saved, so the change here is just to get the code running, and doesn't fix the issue of the log files from regating not being saved.